stream-producer/requirements.txt
boto3==1.7.6
botocore==1.10.6
certifi==2018.4.16
chardet==3.0.4
docutils==0.14
future==0.16.0
idna==2.6
jmespath==0.9.3
oauthlib==2.0.7
python-dateutil==2.7.2
python-twitter==3.4.1
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
wheel==0.24.0
